- Define typography style `displayMedium` for buttons - Update `MyButton` component to use `displayMedium` typography - Organize `ButtonComponent.kt` into `button` package - Remove test button from `Welcome` screen
