+++ app/src/main/java/ar/edu/ort/tp3_ort_2025_parcial/screen/welcome/Welcome.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import ar.edu.ort.tp3_ort_2025_parcial.component.MyButton
+import ar.edu.ort.tp3_ort_2025_parcial.component.button.MyButton
 
 @Composable
 fun Welcome(
@@ -29,10 +29,6 @@ fun Welcome(
             Text(
                 text = "Welcome Screen",
                 fontSize = 30.sp,
-            )
-            MyButton(
-                onClick = {},
-                text = "Test"
             )
         }
     }
