+++ app/src/main/java/ar/edu/ort/tp3_ort_2025_parcial/ui/theme/Type.kt
@@ -44,4 +44,13 @@ val Typography = Typography(
         lineHeight = 19.2.sp,
         letterSpacing = 0.5.sp
     ),
+    // Boton
+    displayMedium = TextStyle(
+        fontSize = 16.sp,
+        lineHeight = 28.8.sp,
+        fontFamily = FontFamily(Font(R.font.poppins_bold)),
+        fontWeight = FontWeight(700),
+        color = White,
+        textAlign = TextAlign.Center,
+    )
     )
